BUYER_SELLER/chat/requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
python-dotenv
streamlit
langgraph
langchain-openai
langchain-core